Bugfix - Barrel Service in Constructor
I kept getting the error 'constructor Can't resolve all parameters for'
when I would add AddressService to the AddressDialog component.  As per
this article
(http://stackoverflow.com/questions/37997824/angular-2-di-error-exception-cant-resolve-all-parameters),
it turns out to be an issue with using Barrels whereby a circular
dependency is created.  The resolution was to directly declare the
Service as opposed to declaring it from the Barrel.