Bugfix Fast Json Patch Typescript Issues
the fast-json-patch package would throw typescript compiler errors and I
didn't have intellisense for the library.
First I was able to fix the SystemJs Module loader issue with having to
name the script being used in the component.
I then was able to get typings installed for it  as well as upgrade it
to version 1.1.4