Updated Automapper Map for companies and upgraded Automapper
It was incorrect to map <task<list<Company>>> which was why I was
receiving the automapper exception.  Also the nature of the mapping that
I was doing (ie ProjectTo that worked with automapper 6.0.2) was issuing
multiple queries and then filtering the mapped results of those queries.
By changing the syntax to this, there is one query that is issued and
returned.